+++ example/example-view-camera.js
@@ -1,0 +1,1 @@
+// To be documented
