--- ios/Resources/README.md
@@ -1,10 +1,0 @@
-Files in this folder are copied directory into the compiled product directory
-when the iOS app is compiled:
-
-    <project-dir>/build/iphone/build/Products/<Debug-iphonesimulator|Release-iphoneos>/<app name>.app/
-
-Place your module's iOS bundles and localization files in this folder.
-
-Files in this directory are copied directly on top of whatever files are already
-in the build directory, so please be careful that your files don't clobber
-essential project files or files from other modules.
